+++ sources/server/Server.ts
@@ -124,6 +124,8 @@ class Server extends events.EventEmitter implements Core.Server {
             this.on("checkExpectation", this.__requestCallback.bind(this));
         }
 
+        this._server.setTimeout(30000);
+
         this._server.keepAliveTimeout = this._keepAlive;
 
         this._server.listen(
@@ -177,6 +179,15 @@ class Server extends events.EventEmitter implements Core.Server {
 
             request.host = this._host;
         }
+
+        request.on("aborted", function(this: Core.ServerRequest) {
+
+            this.aborted = true;
+
+        }).on("close", function(this: Core.ServerRequest) {
+
+            this.closed = true;
+        });
     }
 
     protected async __requestCallback(